Add CNAME file back in
Page stopped loading after last update and it may be because this file was automatically deleted upon build. This may be needed in the docs folder in order the DNS to work correctly. Committing to test this out. If correct, will add automation to ensure this file remains in the docs folder.